fix: clean up entry points
